# prometheusremotewriteexporter: Remove unnecessary buffer copy in proto conversion