This happens because the proto message is generated with gogoproto, and it will go on the path https://github.com/gogo/protobuf/blob/f67b8970b736e53dbd7d0a27146c8f1ac52f74e5/proto/table_marshal.go#L2977

Benchmark numbers show improvement as well.

**After:**
```
BenchmarkExecute/numSample=100
BenchmarkExecute/numSample=100-12         	   21752	     52500 ns/op	    6635 B/op	      83 allocs/op
BenchmarkExecute/numSample=1000
BenchmarkExecute/numSample=1000-12        	   10000	    131918 ns/op	   23096 B/op	      86 allocs/op
BenchmarkExecute/numSample=10000
BenchmarkExecute/numSample=10000-12       	    1479	    886487 ns/op	   41477 B/op	      86 allocs/op
```

**Before:**
```
BenchmarkExecute/numSample=100
BenchmarkExecute/numSample=100-12         	   20642	     55670 ns/op	    6635 B/op	      83 allocs/op
BenchmarkExecute/numSample=1000
BenchmarkExecute/numSample=1000-12        	    8640	    145875 ns/op	   23121 B/op	      86 allocs/op
BenchmarkExecute/numSample=10000
BenchmarkExecute/numSample=10000-12       	    1015	   1257688 ns/op	  304366 B/op	      87 allocs/op
```