@mx-psi 2025-09-04
This seems like the most likely cause of #42414. Should we revert until we can fix that?