@jmichalek132 2025-09-04
> This seems like the most likely cause of #42414. Should we revert until we can fix that?

I dropped your question into [slack](https://cloud-native.slack.com/archives/C01LSCJBXDZ/p1757001728254799) to get feedback from the maintainers of the prom components.