@jmichalek132 2025-09-04
> This seems like the most likely cause of #42414. Should we revert until we can fix that?

I think that's a good idea but, I am not an approver for the component so maybe you might want to wait for a reply from them?

cc @ywwg @krajorama @ArthurSens @dashpole