@ArthurSens 2025-09-04
I'm sorry, but I couldn't find the time to review this. If you feel like reverting this one would solve the flaky tests, please go for it. If we notice that the flakiness is still ongoing after the revert, we can also revert the revert.

In an ideal world, merging PRs without approval from codeowners shouldn't happen. However, I feel like we have more PRs being opened than our current group of codeowners can handle, so I don't want to block PRs from collector maintainers because of our slow responsiveness.

I'm open to ideas here