@jmichalek132 2025-09-04
fyi raised the PR to revert it https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/42501